requirements.txt
amqp==2.5.2
anyjson==0.3.3
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
billiard==3.5.0.5
black==19.10b0
celery==4.1.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
defusedxml==0.6.0
Django==3.0.2
django-allauth==0.41.0
django-crispy-forms==1.8.1
idna==2.8
importlib-metadata==1.5.0
kombu==4.6.7
oauthlib==3.1.0
pathspec==0.7.0
Pillow==7.0.0
pkg-resources==0.0.0
psycopg2==2.8.4
python3-openid==3.1.0
pytz==2019.3
redis==3.4.1
regex==2020.1.8
requests==2.22.0
requests-oauthlib==1.3.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.7
vine==1.3.0
zipp==2.1.0
